refactor: extract BLE transport logic into standalone ble.py module
Move all Bluetooth Low Energy logic out of ZendureDevice and ZendureLegacy
into a new ble.py module as plain functions that take an explicit device
argument. This eliminates BLE concerns from the device class hierarchy and
makes the transport layer independently testable.

- Add ble.py with: ble_mac, ble_sources, ble_device_from_source,
  ble_adapter_options, selected_ble_source, ble_mqtt, ble_command,
  discover_ble_mac, SF_COMMAND_CHAR constant
- device.py: remove 9 BLE methods/properties and bleak imports;
  update mqttSelect, bleAdapterSelect, button_press to call ble_transport.*
- manager.py: remove local isBleDevice function; update BLE discovery to use
  ble_transport.ble_mac() and ble_transport.discover_ble_mac()
